+++ tests/Text.Format.tests.js
@@ -32,6 +32,7 @@ tools.test( function identity() {
   expect(upright).equalsTextFormat(ha.left, va.baseline, rac.Angle.up);
   expect(upright).equalsTextFormat(ha.left, va.baseline, 3/4);
 
+  // RELEASE-TODO: test with padding values
   // Assertion Inequality
   expect(centered).not.equalsTextFormat(ha.left,   va.center, 0,   null,  null);
   expect(centered).not.equalsTextFormat(ha.center, va.top,    0,   null,  null);
@@ -84,18 +85,25 @@ tools.test( function identity() {
 });
 
 
+// RELEASE-TODO: test with padding values
 tools.test( function toString() {
   const format = rac.Text.Format(ha.right, va.bottom, 0.12345, 'monospace', 14);
-  expect(format.toString()) .toBe('Text.Format(ha:right va:bottom a:0.12345 f:"monospace" s:14)');
-  expect(format.toString(2)).toBe('Text.Format(ha:right va:bottom a:0.12 f:"monospace" s:14.00)');
+  expect(format.toString())
+    .toBe('Text.Format(ha:right va:bottom a:0.12345 f:"monospace" s:14 p:(0,0))');
+  expect(format.toString(2))
+    .toBe('Text.Format(ha:right va:bottom a:0.12 f:"monospace" s:14.00 p:(0.00,0.00))');
 
   const zeroFormat = rac.Text.Format(ha.right, va.bottom, 0.12345, '', 0);
-  expect(zeroFormat.toString()) .toBe('Text.Format(ha:right va:bottom a:0.12345 f:"" s:0)');
-  expect(zeroFormat.toString(2)).toBe('Text.Format(ha:right va:bottom a:0.12 f:"" s:0.00)');
+  expect(zeroFormat.toString())
+    .toBe('Text.Format(ha:right va:bottom a:0.12345 f:"" s:0 p:(0,0))');
+  expect(zeroFormat.toString(2))
+    .toBe('Text.Format(ha:right va:bottom a:0.12 f:"" s:0.00 p:(0.00,0.00))');
 
   const nullFormat = rac.Text.Format(ha.right, va.bottom, 0.12345, null, null);
-  expect(nullFormat.toString()) .toBe('Text.Format(ha:right va:bottom a:0.12345 f:null s:null)');
-  expect(nullFormat.toString(2)).toBe('Text.Format(ha:right va:bottom a:0.12 f:null s:null)');
+  expect(nullFormat.toString())
+    .toBe('Text.Format(ha:right va:bottom a:0.12345 f:null s:null p:(0,0))');
+  expect(nullFormat.toString(2))
+    .toBe('Text.Format(ha:right va:bottom a:0.12 f:null s:null p:(0.00,0.00))');
 });