+++ changelog.md
@@ -6,7 +6,9 @@ Changelog
 + Overall documentation improvements and corrections
 + Add `ray.text`, `segment.text`, `arc.text`
 + Add `arc.startTangentRay`, `arc.endTangentRay`, `arc.startRadiusSegment`, `arc.endRadiusSegment`
-+ Add `text.reverse`, `text.upright`
++ Add `text.reverse`, `text.upright`, `text.withPaddings`
++ Add properties `textFormat.hPadding`, `textFormat.vPadding`
++ Add `textFormat.withPaddings`
 + Add `rac.Text.Format.bottomCenter`, `rac.Text.Format.bottomRight`
 + Add `rac.Text.Format.baseline...` constants
 + Add abbreviated shorthands for `rac.Text.Format` ready-made formats, like `rac.Text.Format.tl`